Fixed progress normalized  (returned 0 because integer / integer..) Made EventProgressData `Default` static method to make it easier to make EventProgressData without the parsing functionality.

+++ Assets/Scripts/Internal/Timeline/BaseTimelineEventProgressor.cs
@@ -42,7 +42,7 @@ public abstract class BaseTimelineEventProgressor
 			if(GoalValue == 0)
 				return 0f;
 
-			return CurrentValue / GoalValue;
+			return (float)CurrentValue / GoalValue;
 		}
 	}
